feat: continued with import and changes after review
