+++ app/views/groups/show.html.erb
@@ -6,7 +6,7 @@
   (CommonComponents::ButtonComponent.new(label: t(:edit_group), href: edit_group_path(@group)) if policy(@group).edit?),
   (CommonComponents::ButtonComponent.new(label: t(:generate_report), href: reports_group_path(@group), options: { target: '_blank' } ) if @students.length > 0),
   (CommonComponents::ButtonComponent.new(label: t(:export_students_csv), href: students_url(group_id: @group.id, format: :csv)) if @students.length > 0),
-  (CommonComponents::ButtonComponent.new(label: 'Import Students', href: import_group_path(@group), options: { data: { turbo_stream: true } } ) if policy(@group).import_students?),
+  (CommonComponents::ButtonComponent.new(label: 'Import Student CSV', href: import_group_students_url(@group), options: { data: { turbo_stream: true } } ) if policy(@group).import_students?),
   (CommonComponents::FormButtonComponent.new(label: t(:restore_group), path: undelete_group_path(@group), method: :post) if policy(@group).destroy? && @group.deleted_at),
   (CommonComponents::DangerousFormButtonComponent.new(label: t(:delete_group), path: group_path(@group), method: :delete) if policy(@group).destroy? && !@group.deleted_at)
 ].compact)  do |h|
