+++ config/routes.rb
@@ -54,11 +54,11 @@ Rails.application.routes.draw do
 
   resources :groups, only: %i[index new create show edit update destroy] do
     member { post :undelete }
-    member { post :import_students }
-    member { get :import }
 
     resources :students, controller: :group_students, only: %i[new create edit update] do
       member { post :cancel_edit }
+      get :import, on: :collection
+      post :import_students, on: :collection
     end
   end
